Finished AI
Finished the mech behaviour tree. Added in attack aninations and the death strate which will cancel all AI functions. Allowed the AI to be able to sense sounds and go to the location Changed the gun model from the placeholder. Added in some code to make the map more interactive.